requirements.txt
Flask==2.3.2
paho-mqtt==1.6.1
python-dotenv==1.0.0
gunicorn==20.1.0
Werkzeug==2.3.7